main/requirements.txt
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.4.2
numexpr==2.7.3
numpy==1.19.5
pandas==1.2.5
pandastable==0.12.2.post1
Pillow==8.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
xlrd==2.0.1
